refactor(overlayAdmin): split POST handlers into overlayAdminMutations.ts
overlayAdmin.ts now owns only the GET /settings view handler, the
KNOWN_ERRORS/KNOWN_SUCCESSES sets, and fetchTwitchRewards. All four POST
handlers and their helpers (requireStreamer, toStringArray, parseWeight,
multer setup) move to overlayAdminMutations.ts. overlayAdmin.ts imports
and mounts the mutations router, so server.ts and the test file are
unchanged.

Addresses the high-complexity qlty finding raised on this PR.
